remove qq related from core and sdk
