+++ autopcr/core/sdkclient.py
@@ -1,7 +1,7 @@
 from enum import Enum
 from typing import Tuple
 from abc import abstractmethod
-from ..sdk.validator import Validator
+from ..sdk.validator import remoteValidator
 from copy import deepcopy
 from ..constants import DEFAULT_HEADERS, IOS_HEADERS
 
@@ -11,12 +11,10 @@ class platform(Enum):
 
 class account:
     type: platform
-    qq: str
     username: str
     password: str
     
-    def __init__(self, qid: str, usr: str, pwd: str, type: platform):
-        self.qq = qid
+    def __init__(self, usr: str, pwd: str, type: platform):
         self.username = usr
         self.password = pwd
         self.type = type
@@ -26,7 +24,10 @@ async def _defaultLogger(msg):
 
 class sdkclient:
     
-    def __init__(self, info: account, captchaVerifier=Validator, errlogger=_defaultLogger):
+    async def _default_post_login(self):
+        pass
+
+    def __init__(self, info: account, captchaVerifier=remoteValidator, errlogger=_defaultLogger):
         self.captchaVerifier = captchaVerifier
         self.errlogger = errlogger
         if info.type == platform.Android:
@@ -36,14 +37,19 @@ class sdkclient:
         else:
             raise ValueError(f"Invalid platform {info.type}")
         self._account = info
+        self.post_login = self._default_post_login
     '''
     returns: uid, access_key
     '''
     @abstractmethod
     async def login(self) -> Tuple[str, str]: ...
 
+    async def invoke_post_login(self):
+        await self.post_login()
+        self.post_login = self._default_post_login
+    
     async def do_captcha(self):                                
-        return await self.captchaVerifier(self.qq)
+        return await self.captchaVerifier(self)
     
     def header(self):
         if self._account.type == platform.Android:
@@ -76,10 +82,6 @@ class sdkclient:
     def account(self):
         return self._account.username
 
-    @property
-    def qq(self):
-        return self._account.qq
-    
     @property
     @abstractmethod
     def reskey(self) -> str: ...
