+++ autopcr/sdk/sdkclients.py
@@ -7,7 +7,6 @@ class bsdkclient(sdkclient):
     async def login(self):
         while True:
             resp = await login(
-                self._account.qq,
                 self._account.username,
                 self._account.password,
                 self.captchaVerifier
